Password_Manager_with_mysql/requirements.txt
mysql_connector_repackaged==0.3.1
pycryptodome==3.21.0
pyperclip==1.9.0
rich==13.9.4
